requirements_dev.txt
pip==10.0.1
bumpversion==0.5.3
watchdog==0.8.3
flake8==3.5.0
tox==2.9.1
coverage==4.5.1
Sphinx==1.7.1
sphinxcontrib-httpdomain==1.7.0
sphinx-rtd-theme==0.4.0

pytest==3.4.2
pytest-runner==2.11.1

Django==2.0.7
django-livesync==0.4
django-constance==2.2.0
django-picklefield==1.0.0
djangorestframework==3.8.2
kubernetes==6.0.0

